[FIX] l10n_it_edi: customer outside EU
- Create a partner outside Europe
- Set a VAT number
- Create an invoice for the partner
- Post the invoice

The `IdPaese` and `IdCodice` is obtained from the VAT number, but it is
not correct for partners outside Europe: the VAT number should always be
`OO99999999999`.

A workaround is to set the VAT number of the partner to
`XXOO99999999999`, where `XX` is the country code. However, in case of
multi-company with shared partners, another company might need the
proper VAT number.

In case of a customer outside EU, we:
- get the `IdPaese` from the country of the partner
- set the `IdCodice` to `OO99999999999`

We also add the `IdPaese` to customers without VAT.

opw-2355842

closes odoo/odoo#60416

X-original-commit: 35267c59c32f747351e8741cfe6011749914e809
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>