[FIX] mass_mailing: restore button text colors
The button text colors were forced to the default link color while that
color should only be forced for links which are not buttons.

Note: this fix will not impact existing mail templates thanks to
transcoding. That's why it is possible to fix safely in stable versions.

Part of https://github.com/odoo/odoo/pull/60308
opw-2360756

closes odoo/odoo#60308

X-original-commit: 3d8c6eca5ffa657485bdf4bec84c07b4f978a71d
Related: odoo/enterprise#14217
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>