[FIX] web_editor: restore overlay top options on top of the page
When the overlay covers a snippet, if that snippet is at the top of
the page, the options at the top of the overlay are hidden. A system
was made to make them change position in that case but this was broken
by [1], leaving the toggling of a class for that case but without its
related scss.

This commit restores that scss and also improves the detection of that
problematic case.

[1]: https://github.com/odoo/odoo/commit/66cb7a4f1cc4febdd2d97107c5f5b1d9d97d71f7

closes odoo/odoo#60467

X-original-commit: 801c467f33ccc5affe2ce70e9f7dd86cf83bf570
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>