[FIX] mail: Fix the issue with activity widget dropdown scrollbar
This commit fixes the issue with dropdown of activity in kanban.
The activity dropdown was cut with document inspectorbar in document
kanban view due to data-boundry window on dropdown.

Note:
Dropdown of list activity was cut on the list view and to fix that
commit[1] was done but it has side effect on document kanban view.
another fix was done in commit[2] so now commit[1] is not useful so
here revert that commit.

[1] - https://github.com/odoo/odoo/commit/6f553fd1f75aa38265e285b45854f691d14aef23
[2] - https://github.com/odoo/odoo/commit/ec53889368d579442a3eb40a99c66e76f9a52a35

closes odoo/odoo#61895

Taskid: 2336146
X-original-commit: 6b32689b00e96d9cfc165fe31a3a121d97874321
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>